+++ lib/simply_stored/couch/has_one.rb
@@ -64,7 +64,7 @@ module SimplyStored
           options = {
             :dependent => :nullify,
             :class_name => name.to_s.singularize.camelize,
-            :foreign_key => owner_clazz.name.singularize.underscore.foreign_key
+            :foreign_key => nil
           }.update(options)
           @name, @options = name, options